refactor(frontend): source DIRECTUS_URL from the endpoints module
loadPlaylist.ts was the de-facto shared constant, imported by ten
modules that have nothing to do with playlists.
